+++ py/projects/functional_web_tests/tests/config/locators.py
@@ -6,5 +6,6 @@ class locator(Locator):
     logo = ('id', 'site-logo')
     go_to_login = ('id', 'login')
     login_user = ('id', 'login_user')
+    login_email = ('id', 'login_email')
     login_pass = ('id', 'login_pass')
     login_button = ('id', 'login_button')
