+++ py/projects/functional_web_tests/tests/conftest.py
@@ -1,3 +1,20 @@
 import pytest
 from tests.services.app import App
 from tests.config.supported_platforms import platforms
+
+
+@pytest.fixture(scope='session', params=platforms.get())
+def app(request):
+    """fixture for cross-platform model-based application
+    always use session scope for this type of fixture & parameterize your platforms
+    :rtype: App
+    """
+    if isinstance(request.param, list) or isinstance(request.param, tuple):
+        APPS = []
+        for platform in request.param:
+            APPS.append(App.from_platform(platform, request))
+            request.addfinalizer(APPS[-1].driver.quit)
+        return APPS
+    APP = App.from_platform(request.param, request)
+    request.addfinalizer(APP.driver.quit)
+    return APP
