+++ py/projects/functional_web_tests/tests/pages/components/login.py
@@ -4,6 +4,9 @@ from tests.config.locators import locator
 
 class Login(page):
     """login component"""
+    def login_email(self):
+        return self.find(locator.login_email)
+
     def login_user(self):
         return self.find(locator.login_user)
 
@@ -26,3 +29,10 @@ class Login(page):
         self.login_button().click()
         self.wait_until_page_loads()
         self.find(locator.text(error))
+
+    def with_email(self, email, password):
+        self.login_email().send(email)
+        self.login_pass().send(password)
+        self.login_button().click()
+        self.wait_until_page_loads()
+        self.confirm_element_is_deleted(locator.login_email)
